Reform of access rights
We try to bring the model of access rights closer to Django's built-in
permission model.

  - The "trans.upload_translation_pdf" right is now used consistently
    and it is defined in the DB model, so that it's offered by the
    admin interface. By default, it's not assigned to anybody.

  - A new right "trans.send_notifications" has been added. It controls
    who is permitted to send notifications to all users. By default,
    it's not assigned to anybody, so only admin has is.

  - Deletion of all notifications now requires delete rights on the
    table of notifications. Previously, it was allowed to everybody
    whose Django user had the is_staff attribute set, which includes
    ISC and printing staff.

  - When a request is rejected because of insufficient access right,
    HTTP status 403 is returned instead of redirecting the user to their
    home page.

+++ trans/views/notification.py
@@ -1,12 +1,16 @@
+from django.core.exceptions import PermissionDenied
 from django.http.response import HttpResponseRedirect
 from django.shortcuts import render
 from django.views.generic import View
 from django.http import JsonResponse, HttpResponseNotFound, HttpResponseBadRequest, HttpResponse
 from django.contrib.auth.mixins import LoginRequiredMixin
+import logging
 
 from trans.models import Notification, User
 from trans.utils.notification import get_all_notifs, read_this_notif, read_all_notifs, reset_notification_cache
-from trans.views.admin import StaffCheckMixin
+
+
+logger = logging.getLogger(__name__)
 
 
 class ReadNotifications(LoginRequiredMixin, View):
@@ -34,18 +38,24 @@ class ReadNotifications(LoginRequiredMixin, View):
         return HttpResponse("Success")
 
 
-class SendNotification(StaffCheckMixin, View):
+class SendNotification(View):
     def post(self, request):
+        if not request.user.has_perm('trans.send_notifications'):
+            logger.warning(f'Missing permission trans.send_notifications for user {request.user.username}')
+            raise PermissionDenied('Notification rights required')
+
         title = request.POST['title']
         description = request.POST['description']
         notif = Notification(title=title, description=description)
         notif.save()
         return HttpResponseRedirect(request.META['HTTP_REFERER'])
 
-from django.contrib.admin.views.decorators import staff_member_required
 
-@staff_member_required
 def reset_notifications(request):
+    if not request.user.has_perm('trans.delete_notification'):
+        logger.warning(f'Missing permission trans.send_notifications for user {request.user.username}')
+        raise PermissionDenied('Notification rights required')
+
     Notification.objects.all().delete()
     reset_notification_cache(User.objects.all())
     return HttpResponseRedirect(request.META["HTTP_REFERER"])
